Fix `OpenALWrapper` wrong assumption of object 0
In OpenAL, object handle 0 is not a null object (see here:
https://stackoverflow.com/questions/71095893/can-an-openal-source-ever-be-0)

This commit removes that assumption by wrapping the handle in an
`std::optional<>` and adjust the code accordingly. The lint
`bugprone-unchecked-optional-access` did come up because I was
unwraping an the unchecked optional wrapper, so this API may be somewhat
unsafe to use, in order to stay close with original API of `unique_ptr`
and `optional`. I also provided an `isValid` method that acted somewhat
like `unique_ptr`'s `operator bool` (or comparison with nullptr) and
`optional`'s `has_value()/operator bool()`